Move server into own module
